% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stattotex.R
\name{stattotex}
\alias{stattotex}
\title{stattotex}
\usage{
stattotex(variable_name, variable_value, filename, clear_file = FALSE)
}
\arguments{
\item{variable_name}{Name of the variable to be saved (string)}

\item{variable_value}{Value of the variable to be saved}

\item{filename}{Name of the file to save the variable to (string)}

\item{clear_file}{If TRUE, clears the contents of the file before writing (default is FALSE)}
}
\description{
Function that takes a variable value and associated name from R and saves into a file that allows for easy-read-in and later formatting in LaTeX
}
\examples{
\dontrun{stattotex(variable_name = "AvgMonthlySales", 
                   variable_value = 1500, 
                   filename = "demo/demoNums.tex", 
                   clear_file = FALSE)}
}
