% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stattotex.R
\name{stattotex}
\alias{stattotex}
\title{stattotex}
\usage{
stattotex(number, number_name, filename)
}
\arguments{
\item{number}{A numeric value to be saved.}

\item{number_name}{A string representing the name/command to be associated with the number.}

\item{filename}{The name of the file where the LaTeX command will be saved.}
}
\description{
Function that takes number and associated name from R and saves into a file that allows for easy-read-in and later formatting in LaTeX
}
\examples{
\dontrun{stattotex("AvgMonthlySales", 1500, "demo/demoNums.tex")}
}
